feat(cli): link a pull request on worktree create and set with --pr
`orca worktree create` gains `--pr <number>` and `orca worktree set` gains
`--pr <number|null>`, wired to the existing `linkedPR` worktree field that
the `worktree.create` / `worktree.set` RPC methods already accept. Until now
the CLI only exposed `--issue` and `--linear-issue`, so review automations
either left the PR link empty or misfiled the PR number as an issue.

The flag mirrors `--issue`: same numeric parsing, `null` clears on set and is
rejected on create. Specs, `--help`, root help, and the CLI docs page mention
it. The issue/PR flag parsing and the result-warning printers move into small
sibling modules, and the flag help map into `flag-help-text.ts`, to keep
`handlers/worktree.ts` and `help.ts` under the max-lines limit.
