+++ docs/site/content/docs/cli/overview.mdx
@@ -38,6 +38,7 @@ orca status --json
 ```
 orca worktree ps --json
 orca worktree create --repo id:<repoId> --name my-task --issue 123 --json
+orca worktree create --repo id:<repoId> --name review-pr-123 --pr 123 --json
 orca worktree current --json
 orca worktree set --worktree active --comment "reproduced bug" --json
 orca worktree rm --worktree id:<id> --force --json
