+++ src/cli/root-help-text-secondary.ts
@@ -48,10 +48,10 @@ export const ROOT_HELP_TEXT_SECONDARY = [
   '  orca environment show --environment <selector> [--json]',
   '  orca environment rm --environment <selector> [--json]',
   '  orca worktree list [--repo <selector>] [--limit <n>] [--json]',
-  '  orca worktree create --name <name> [--repo <selector>|--project <id> [--host <host-id>]|--project-host-setup <id>] [--agent <id>] [--prompt <text>] [--setup run|skip|inherit] [--base-branch <ref>] [--issue <number>] [--linear-issue <identifier-or-url>] [--comment <text>] [--parent-worktree <selector>] [--no-parent] [--run-hooks] [--activate] [--json]',
+  '  orca worktree create --name <name> [--repo <selector>|--project <id> [--host <host-id>]|--project-host-setup <id>] [--agent <id>] [--prompt <text>] [--setup run|skip|inherit] [--base-branch <ref>] [--issue <number>] [--pr <number>] [--linear-issue <identifier-or-url>] [--comment <text>] [--parent-worktree <selector>] [--no-parent] [--run-hooks] [--activate] [--json]',
   '  orca worktree show --worktree <selector> [--json]',
   '  orca worktree current [--json]',
-  '  orca worktree set --worktree <selector> [--display-name <name>] [--issue <number|null>] [--linear-issue <identifier-or-url|null>] [--comment <text>] [--workspace-status <id>] [--parent-worktree <selector>|--no-parent] [--json]',
+  '  orca worktree set --worktree <selector> [--display-name <name>] [--issue <number|null>] [--pr <number|null>] [--linear-issue <identifier-or-url|null>] [--comment <text>] [--workspace-status <id>] [--parent-worktree <selector>|--no-parent] [--json]',
   '  orca worktree rm --worktree <selector> [--force] [--run-hooks] [--json]',
   '  orca worktree ps [--limit <n>] [--json]',
   '  orca file open <path> [--worktree <selector>] [--json]',
@@ -157,6 +157,7 @@ export const ROOT_HELP_TEXT_SECONDARY = [
   '  $ orca repo list',
   '  $ orca worktree create --name agent-task --agent codex --prompt "hi"',
   '  $ orca worktree create --repo name:orca --name cli-test-1 --issue 273',
+  '  $ orca worktree create --repo name:orca --name review-pr-123 --pr 123',
   '  $ orca worktree create --repo name:orca --name linear-task --linear-issue https://linear.app/stably/issue/STA-335/test-issue',
   '  $ orca worktree create --name linear-task --linear-issue STA-335',
   '  $ orca worktree show --worktree branch:Jinwoo-H/cli',
